raycast

+++ engine/src/Core/Time.hpp
@@ -4,15 +4,15 @@ namespace Engine {
 
 class Time {
   private:
-    float m_deltaTime = 0;
-    float m_totalTime = 0;
-    float m_LastFrameTime = 0;
+    double m_deltaTime = 0;
+    double m_totalTime = 0;
+    double m_LastFrameTime = 0;
 
   public:
     void tick();
 
-    float getDeltaSeconds() const { return m_deltaTime; }
-    float getTotalSeconds() const { return m_totalTime; }
+    double getDeltaSeconds() const { return m_deltaTime; }
+    double getTotalSeconds() const { return m_totalTime; }
 };
 
 } // namespace Engine
